First pass at making interra_api into dkan_lunr.
