+++ modules/custom/dkan_lunr/dkan_lunr.routing.yml
@@ -1,0 +1,7 @@
+dkan_lunr.search:
+  path: '/api/v1/search-index.json'
+  defaults:
+   { _controller: '\Drupal\interra_api\Controller\ApiController::search'}
+  requirements:
+    _permission: 'access content'
+
