--- modules/custom/interra_api/interra_api.info.yml
@@ -1,10 +1,0 @@
-name: Interra API
-description: 'Interra API.'
-type: module
-core: 8.x
-dependencies:
-  - dkan_common
-  - dkan_data
-  - dkan_datastore
-  - dkan_schema
-package: DKAN
