--- modules/custom/interra_api/interra_api.routing.yml
@@ -1,30 +1,0 @@
-interra_api.schema:
-  path: '/api/v1/schema.json'
-  defaults:
-   { _controller: '\Drupal\interra_api\Controller\ApiController::schemas'}
-  requirements:
-    _permission: 'access content'
-interra_api.schema_single:
-  path: '/api/v1/schema/{schema_name}'
-  defaults:
-    { _controller: '\Drupal\interra_api\Controller\ApiController::schema'}
-  requirements:
-    _permission: 'access content'
-interra_api.search:
-  path: '/api/v1/search-index.json'
-  defaults:
-   { _controller: '\Drupal\interra_api\Controller\ApiController::search'}
-  requirements:
-    _permission: 'access content'
-interra_api.collecion:
-  path: '/api/v1/collections/{collection}'
-  defaults:
-   { _controller: '\Drupal\interra_api\Controller\ApiController::collection'}
-  requirements:
-    _permission: 'access content'
-interra_api.doc:
-  path: '/api/v1/collections/{collection}/{doc}'
-  defaults:
-   { _controller: '\Drupal\interra_api\Controller\ApiController::doc'}
-  requirements:
-    _permission: 'access content'
