--- modules/custom/interra_api/interra_api.services.yml
@@ -1,7 +1,0 @@
-services:
-    interra_api.service.dataset_modifier:
-        class: Drupal\interra_api\Service\DatasetModifier
-    interra_api.search:
-        class: Drupal\interra_api\Search
-        shared: false
-
